+++ ydb/core/kqp/opt/rbo/kqp_rbo_transformer.cpp
@@ -10,990 +10,6 @@ using namespace NKikimr::NKqp;
 using namespace NYql::NDq;
 
 namespace {
-struct TJoinTableAliases {
-    THashSet<TString> LeftSideAliases;
-    THashSet<TString> RightSideAliases;
-};
-
-struct TAggregationTraits {
-    TVector<TExprNode::TPtr> AggTraitsList;
-    TVector<TInfoUnit> KeyColumns;
-};
-
-THashSet<TString> SupportedAggregationFunctions{"sum", "min", "max", "count"};
-ui64 KqpUniqueAggColumnId{0};
-
-TJoinTableAliases GatherJoinAliasesLeftSideMultiInputs(const TVector<TInfoUnit> &joinKeys, const THashSet<TString> &processedInputs) {
-    TJoinTableAliases joinAliases;
-    for (const auto &joinKey : joinKeys) {
-        if (processedInputs.count(joinKey.Alias)) {
-            joinAliases.LeftSideAliases.insert(joinKey.Alias);
-        } else {
-            joinAliases.RightSideAliases.insert(joinKey.Alias);
-        }
-    }
-    Y_ENSURE(joinAliases.LeftSideAliases.size(), "Left side of the join inputs are empty");
-    Y_ENSURE(joinAliases.RightSideAliases.size() == 1, "Right side of the join should have only one input");
-    return joinAliases;
-}
-
-TJoinTableAliases GatherJoinAliasesTwoInputs(const TVector<TInfoUnit> &joinKeys) {
-    TJoinTableAliases joinAliases;
-    for (ui32 i = 0; i < joinKeys.size(); i += 2) {
-        joinAliases.LeftSideAliases.insert(joinKeys[i].Alias);
-        joinAliases.RightSideAliases.insert(joinKeys[i + 1].Alias);
-    }
-
-    Y_ENSURE(joinAliases.LeftSideAliases.size() == 1, "Left side of the join should have only one input");
-    Y_ENSURE(joinAliases.RightSideAliases.size() == 1, "Right side of the join should have only one input");
-    return joinAliases;
-}
-
-TExprNode::TPtr BuildJoinKeys(const TVector<TInfoUnit> &joinKeys, const TJoinTableAliases &joinAliases, THashSet<TString> &processedInputs,
-                              TExprContext &ctx, TPositionHandle pos) {
-    Y_ENSURE(joinKeys.size() >= 2 && !(joinKeys.size() & 1), "Invalid join key size");
-    TVector<TDqJoinKeyTuple> keys;
-    for (ui32 i = 0; i < joinKeys.size(); i += 2) {
-        auto leftSideKey = joinKeys[i];
-        auto rightSideKey = joinKeys[i + 1];
-        if (joinAliases.LeftSideAliases.count(rightSideKey.Alias)) {
-            std::swap(leftSideKey, rightSideKey);
-        }
-        // clang-format off
-        keys.push_back(Build<TDqJoinKeyTuple>(ctx, pos)
-                           .LeftLabel()
-                               .Value(leftSideKey.Alias)
-                           .Build()
-                           .LeftColumn()
-                               .Value(leftSideKey.ColumnName)
-                           .Build()
-                           .RightLabel()
-                               .Value(rightSideKey.Alias)
-                           .Build()
-                           .RightColumn()
-                               .Value(rightSideKey.ColumnName)
-                           .Build()
-                      .Done());
-        // clang-format on
-        processedInputs.insert(leftSideKey.Alias);
-        processedInputs.insert(rightSideKey.Alias);
-    }
-    return Build<TDqJoinKeyTupleList>(ctx, pos).Add(keys).Done().Ptr();
-}
-
-TExprNode::TPtr BuildAggregationTraits(const TString& originalColName, const TString& resultColName, const TString& aggFunction,
-                                       const TTypeAnnotationNode* resultType, TExprContext &ctx, TPositionHandle pos) {
-    // clang-format off
-    return Build<TKqpOpAggregationTraits>(ctx, pos)
-        .OriginalColName<TCoAtom>()
-            .Value(originalColName)
-        .Build()
-        .AggregationFunction<TCoAtom>()
-            .Value(aggFunction)
-        .Build()
-        .ResultColName<TCoAtom>()
-            .Value(resultColName)
-        .Build()
-        .AggregationFunctionResultType(ExpandType(pos, *resultType, ctx))
-    .Done().Ptr();
-    // clang-format on
-}
-
-TExprNode::TPtr BuildAggregate(TExprNode::TPtr resultExpr, const TVector<TInfoUnit>& keys, const TVector<TExprNode::TPtr>& aggTraitsList,
-                               bool distinctAll,
-                               TExprContext& ctx, TPositionHandle pos) {
-    TVector<TCoAtom> keyColumns;
-    for (const auto& column : keys) {
-        // clang-format off
-        auto keyColumn = Build<TCoAtom>(ctx, pos)
-            .Value(column.GetFullName())
-        .Done();
-        // clang-format on
-        keyColumns.push_back(keyColumn);
-    }
-
-    // clang-format off
-    return Build<TKqpOpAggregate>(ctx, pos)
-        .Input(resultExpr)
-        .AggregationTraitsList<TKqpOpAggregationTraitsList>()
-            .Add(aggTraitsList)
-        .Build()
-        .KeyColumns<TCoAtomList>()
-            .Add(keyColumns)
-        .Build()
-        .DistinctAll<TCoAtom>()
-            .Value(distinctAll ? "True" : "False")
-        .Build()
-    .Done().Ptr();
-    // clang-format on
-}
-
-void BuildSimpleMapElementLambda(TExprNode::TPtr resultExpr, const TVector<std::pair<TInfoUnit, TInfoUnit>>& renamesMap, TVector<TExprNode::TPtr>& mapElements,
-                                 TExprContext& ctx, TPositionHandle pos) {
-    for (const auto& [colName, rename] : renamesMap) {
-        // clang-format off
-        // Just wrap by member
-        auto lambda = ctx.Builder(pos)
-            .Lambda()
-                .Param("arg")
-                .Callable("Member")
-                    .Arg(0, "arg")
-                    .Atom(1, colName.GetFullName())
-                .Seal()
-        .Seal().Build();
-
-        mapElements.push_back(Build<TKqpOpMapElementLambda>(ctx, pos)
-            .Input(resultExpr)
-            .Variable()
-                .Value(colName.GetFullName())
-            .Build()
-            .Lambda(lambda)
-        .Done().Ptr());
-        // clang-format on
-    }
-}
-
-TExprNode::TPtr BuildExpressionMap(TExprNode::TPtr resultExpr,
-                                   const TVector<std::pair<TInfoUnit, TExprNode::TPtr>>& aggFieldsExpressionsMap,
-                                   const TVector<std::pair<TInfoUnit, TInfoUnit>>& aggFieldsRenamesMap,
-                                   const TVector<std::pair<TInfoUnit, TInfoUnit>>& groupByKeysRenamesMap,
-                                   const THashMap<uint32_t, std::pair<TInfoUnit, TExprNode::TPtr>>& groupByKeysExpressionsMap,
-                                   TExprContext& ctx, TPositionHandle pos) {
-    // Add expressions
-    TVector<TExprNode::TPtr> mapElements;
-    for (const auto& [colName, expr] : aggFieldsExpressionsMap) {
-        // clang-format off
-        mapElements.push_back(Build<TKqpOpMapElementLambda>(ctx, pos)
-            .Input(resultExpr)
-            .Variable()
-                .Value(colName.GetFullName())
-            .Build()
-            .Lambda(expr)
-        .Done().Ptr());
-        // clang-format on
-    }
-
-    // Add other columns and wrap them with member.
-    BuildSimpleMapElementLambda(resultExpr, aggFieldsRenamesMap, mapElements, ctx, pos);
-    BuildSimpleMapElementLambda(resultExpr, groupByKeysRenamesMap, mapElements, ctx, pos);
-
-    // Add expressions for group by keys.
-    for (const auto& [_, pair] : groupByKeysExpressionsMap) {
-        // clang-format off
-        mapElements.push_back(Build<TKqpOpMapElementLambda>(ctx, pos)
-            .Input(resultExpr)
-            .Variable()
-                .Value(pair.first.GetFullName())
-            .Build()
-            .Lambda(pair.second)
-        .Done().Ptr());
-        // clang-format on
-    }
-
-    // clang-format off
-    return Build<TKqpOpMap>(ctx, pos)
-        .Input(resultExpr)
-        .MapElements()
-            .Add(mapElements)
-        .Build()
-        .Project()
-            .Value("true")
-        .Build()
-    .Done().Ptr();
-    // clang-format on
-}
-
-void BuildMapElementRename(TExprNode::TPtr resultExpr, const TVector<std::pair<TInfoUnit, TInfoUnit>>& renamesMap,
-                           TVector<TExprNode::TPtr>& mapElements, TExprContext& ctx, TPositionHandle pos) {
-    for (const auto& [colName, newColName] : renamesMap) {
-        // clang-format off
-        mapElements.push_back(Build<TKqpOpMapElementRename>(ctx, pos)
-            .Input(resultExpr)
-            .Variable()
-                .Value(newColName.GetFullName())
-            .Build()
-            .From()
-                .Value(colName.GetFullName())
-            .Build()
-        .Done().Ptr());
-        // clang-format on
-    }
-}
-
-TExprNode::TPtr BuildRenameMap(TExprNode::TPtr resultExpr, const TVector<std::pair<TInfoUnit, TInfoUnit>>& aggFieldsRenamesMap,
-                               const TVector<std::pair<TInfoUnit, TInfoUnit>>& groupByKeysRenamesMap, TExprContext& ctx, TPositionHandle pos) {
-    TVector<TExprNode::TPtr> mapElements;
-    BuildMapElementRename(resultExpr, aggFieldsRenamesMap, mapElements, ctx, pos);
-    BuildMapElementRename(resultExpr, groupByKeysRenamesMap, mapElements, ctx, pos);
-
-    // clang-format off
-    return Build<TKqpOpMap>(ctx, pos)
-        .Input(resultExpr)
-        .MapElements()
-            .Add(mapElements)
-        .Build()
-        .Project()
-            .Value("false")
-        .Build()
-    .Done().Ptr();
-    // clang-format on
-}
-
-TString GenerateUniqueColumnName(const TString &colName) {
-    TStringBuilder strBuilder;
-    strBuilder << "_kqp_agg_input_";
-    strBuilder << colName;
-    strBuilder << "_";
-    strBuilder << ToString(KqpUniqueAggColumnId++);
-    return strBuilder;
-}
-
-void ToCamelCase(std::string& s) {
-    char previous = ' ';
-    auto f = [&](char current) {
-        char result = (std::isblank(previous) && std::isalpha(current)) ? std::toupper(current) : std::tolower(current);
-        previous = current;
-        return result;
-    };
-    std::transform(s.begin(), s.end(), s.begin(), f);
-}
-
-TExprNode::TPtr GetPgCallable(TExprNode::TPtr input, const TString& callableName) {
-    auto isPgCallable = [&](const TExprNode::TPtr& node) -> bool {
-        if (node->IsCallable(callableName)) {
-            return true;
-        }
-        return false;
-    };
-
-    return FindNode(input, isPgCallable);
-}
-
-TExprNode::TPtr GetAtom(TExprNode::TPtr input, const TString& atomName) {
-    auto isPgAtom = [&](const TExprNode::TPtr& node) -> bool {
-        if (node->IsAtom(atomName)) {
-            return true;
-        }
-        return false;
-    };
-
-    return FindNode(input, isPgAtom);
-}
-
-TExprNode::TPtr ReplacePgOps(TExprNode::TPtr input, TExprContext &ctx) {
-    if (input->IsLambda()) {
-        auto lambda = TCoLambda(input);
-
-        // clang-format off
-            return Build<TCoLambda>(ctx, input->Pos())
-                .Args(lambda.Args())
-                .Body(ReplacePgOps(lambda.Body().Ptr(), ctx))
-            .Done().Ptr();
-        // clang-format on
-    } else if (input->IsCallable("PgAnd")) {
-        // clang-format off
-            return ctx.Builder(input->Pos())
-                .Callable("ToPg")
-                    .Callable(0, "And")
-                        .Callable(0, "FromPg")
-                            .Add(0, ReplacePgOps(input->ChildPtr(0), ctx))
-                        .Seal()
-                        .Callable(1, "FromPg")
-                            .Add(0, ReplacePgOps(input->ChildPtr(1), ctx))
-                        .Seal()
-                    .Seal()
-                .Seal()
-            .Build();
-        // clang-format on
-
-    } else if (input->IsCallable("PgOr")) {
-        // clang-format off
-            return ctx.Builder(input->Pos())
-                .Callable("ToPg")
-                    .Callable(0, "Or")
-                        .Callable(0, "FromPg")
-                            .Add(0, ReplacePgOps(input->ChildPtr(0), ctx))
-                        .Seal()
-                        .Callable(1, "FromPg")
-                            .Add(0, ReplacePgOps(input->ChildPtr(1), ctx))
-                        .Seal()
-                    .Seal()
-                .Seal()
-            .Build();
-            // clnag-format on
-    } else if (input->IsCallable()){
-            TVector<TExprNode::TPtr> newChildren;
-            for (auto c : input->Children()) {
-                newChildren.push_back(ReplacePgOps(c, ctx));
-            }
-            // clang-format off
-            return ctx.Builder(input->Pos())
-                .Callable(input->Content())
-                    .Add(std::move(newChildren))
-                .Seal()
-            .Build();
-        // clang-format on
-    } else if (input->IsList()) {
-        TVector<TExprNode::TPtr> newChildren;
-        for (auto c : input->Children()) {
-            newChildren.push_back(ReplacePgOps(c, ctx));
-        }
-        // clang-format off
-            return ctx.Builder(input->Pos())
-                .List()
-                    .Add(std::move(newChildren))
-                .Seal()
-            .Build();
-        // clang-format on
-    } else {
-        return input;
-    }
-}
-
-TExprNode::TPtr BuildSort(TExprNode::TPtr input, TExprNode::TPtr sort, TExprContext &ctx) {
-    TVector<TExprNode::TPtr> sortElements;
-
-    for (auto sortItem : sort->Child(1)->Children()) {
-        auto sortLambda = sortItem->Child(1);
-        auto direction = sortItem->Child(2);
-        auto nullsFirst = sortItem->Child(3);
-
-        // clang-format off
-        sortElements.push_back(Build<TKqpOpSortElement>(ctx, input->Pos())
-            .Input(input)
-            .Direction(direction)
-            .NullsFirst(nullsFirst)
-            .Lambda(sortLambda)
-            .Done().Ptr());
-        // clang-format on
-    }
-
-    // clang-format off
-    return Build<TKqpOpSort>(ctx, input->Pos())
-        .Input(input)
-        .SortExpressions().Add(sortElements).Build()
-        .Done().Ptr();
-    // clang-format off
-}
-
-TExprNode::TPtr RewritePgSelect(const TExprNode::TPtr &node, TExprContext &ctx, const TTypeAnnotationContext &typeCtx) {
-    Y_UNUSED(typeCtx);
-
-    TVector<TString> finalColumnOrder;
-
-    auto setItems = GetSetting(node->Head(), "set_items")->TailPtr();
-    TVector<TExprNode::TPtr> setItemsResults;
-    for (ui32 i = 0; i < setItems->ChildrenSize(); ++i) {
-        auto setItem = setItems->ChildPtr(i);
-
-        TVector<TExprNode::TPtr> resultElements;
-        // In pg syntax duplicate attributes are allowed in the results, but we need to rename them
-        // We use the counters for this purpose
-        THashMap<TString, int> resultElementCounters;
-
-        TExprNode::TPtr joinExpr;
-        TExprNode::TPtr filterExpr;
-        TExprNode::TPtr lastAlias;
-
-        auto from = GetSetting(setItem->Tail(), "from");
-        THashMap<TString, TExprNode::TPtr> aliasToInputMap;
-        TVector<TExprNode::TPtr> inputsInOrder;
-
-        if (from) {
-            for (auto fromItem : from->Child(1)->Children()) {
-                // From item can be a table read with an alias or a subquery with an alias
-                // In case of a subquery, we have already translated PgSelect of the nested subquery
-                // so we just need to remove TKqpOpRoot and plug in the translated subquery
-
-                auto childExpr = fromItem->ChildPtr(0);
-                auto alias = fromItem->Child(1);
-                TExprNode::TPtr fromExpr;
-
-                if (TKqpOpRoot::Match(childExpr.Get())) {
-                    auto opRoot = TKqpOpRoot(childExpr);
-
-                    TVector<TExprNode::TPtr> subqueryElements;
-
-                    // We need to rename all the IUs in the subquery to reflect the new alias
-                    auto subqueryType = childExpr->GetTypeAnn()->Cast<TListExprType>()->GetItemType()->Cast<TStructExprType>();
-                    for (auto item : subqueryType->GetItems()) {
-                        auto orig = TString(item->GetName());
-                        auto unit = TInfoUnit(orig);
-                        auto renamedUnit = TInfoUnit(TString(alias->Content()), unit.ColumnName);
-
-                        // clang-format off
-                        subqueryElements.push_back(Build<TKqpOpMapElementRename>(ctx, node->Pos())
-                            .Input(opRoot.Input())
-                            .Variable().Value(renamedUnit.GetFullName()).Build()
-                            .From().Value(unit.GetFullName()).Build()
-                        .Done().Ptr());
-                        // clang-format on
-                    }
-
-                    // clang-format off
-                    fromExpr = Build<TKqpOpMap>(ctx, node->Pos())
-                        .Input(opRoot.Input())
-                        .MapElements().Add(subqueryElements).Build()
-                        .Project().Value("true").Build()
-                    .Done().Ptr();
-                    // clang-format on
-                }
-
-                else {
-                    auto readExpr = TKqlReadTableRanges(childExpr);
-
-                    // clang-format off
-                    fromExpr = Build<TKqpOpRead>(ctx, node->Pos())
-                        .Table(readExpr.Table())
-                        .Alias(alias)
-                        .Columns(readExpr.Columns())
-                    .Done().Ptr();
-                    // clang-format on
-                }
-
-                aliasToInputMap.insert({TString(alias->Content()), fromExpr});
-                inputsInOrder.push_back(fromExpr);
-                lastAlias = alias;
-            }
-        }
-
-        THashSet<TString> processedInputs;
-        auto joinOps = GetSetting(setItem->Tail(), "join_ops");
-        if (joinOps) {
-            for (ui32 i = 0; i < joinOps->Tail().ChildrenSize(); ++i) {
-                ui32 tableInputsCount = 0;
-                auto tuple = joinOps->Tail().Child(i);
-                for (ui32 j = 0; j < tuple->ChildrenSize(); ++j) {
-                    auto join = tuple->Child(j);
-                    auto joinType = join->Child(0)->Content();
-                    if (joinType == "push") {
-                        ++tableInputsCount;
-                        continue;
-                    }
-
-                    Y_ENSURE(join->ChildrenSize() > 1 && join->Child(1)->ChildrenSize() > 1);
-                    auto pgResolvedOps = FindNodes(join->Child(1)->Child(1)->TailPtr(), [](const TExprNode::TPtr &node) {
-                        if (node->IsCallable("PgResolvedOp")) {
-                            return true;
-                        } else {
-                            return false;
-                        }
-                    });
-
-                    // FIXME: join on clause may include expressions, we need to handle this case
-                    TVector<TInfoUnit> joinKeys;
-                    for (const auto &pgResolvedOp : pgResolvedOps) {
-                        TVector<TInfoUnit> keys;
-                        GetAllMembers(pgResolvedOp, keys);
-                        joinKeys.insert(joinKeys.end(), keys.begin(), keys.end());
-                    }
-
-                    TJoinTableAliases joinAliases;
-                    TExprNode::TPtr leftInput;
-                    TExprNode::TPtr rightInput;
-
-                    if (tableInputsCount == 2) {
-                        joinAliases = GatherJoinAliasesTwoInputs(joinKeys);
-                        const auto leftSideAlias = *joinAliases.LeftSideAliases.begin();
-                        const auto rightSideAlias = *joinAliases.RightSideAliases.begin();
-                        Y_ENSURE(aliasToInputMap.count(leftSideAlias), "Left side alias is not present in input tables");
-                        Y_ENSURE(aliasToInputMap.count(rightSideAlias), "Right sided alias is not present input tables");
-                        leftInput = aliasToInputMap[leftSideAlias];
-                        rightInput = aliasToInputMap[rightSideAlias];
-                    } else if (tableInputsCount == 1) {
-                        joinAliases = GatherJoinAliasesLeftSideMultiInputs(joinKeys, processedInputs);
-                        const auto rightSideAlias = *joinAliases.RightSideAliases.begin();
-                        Y_ENSURE(aliasToInputMap.contains(rightSideAlias), "Right side alias is not present in input tables");
-                        leftInput = joinExpr;
-                        rightInput = aliasToInputMap[rightSideAlias];
-                    }
-
-                    auto joinKind = TString(joinType);
-                    ToCamelCase(joinKind.MutRef());
-
-                    // clang-format off
-                    joinExpr = Build<TKqpOpJoin>(ctx, node->Pos())
-                        .LeftInput(leftInput)
-                        .RightInput(rightInput)
-                        .JoinKind()
-                            .Value(joinKind)
-                        .Build()
-                        .JoinKeys(BuildJoinKeys(joinKeys, joinAliases, processedInputs, ctx, node->Pos()))
-                    .Done().Ptr();
-                    // clang-format on
-                    tableInputsCount = 0;
-                }
-            }
-
-            // Build in order
-            if (!joinExpr) {
-                ui32 inputIndex = 0;
-                if (inputsInOrder.size() > 1) {
-                    while (inputIndex < inputsInOrder.size()) {
-                        auto leftTableInput = inputIndex == 0 ? inputsInOrder[inputIndex] : joinExpr;
-                        auto rightTableInput = inputIndex == 0 ? inputsInOrder[inputIndex + 1] : inputsInOrder[inputIndex];
-                        auto joinKeys = Build<TDqJoinKeyTupleList>(ctx, node->Pos()).Done();
-                        // clang-format off
-                        joinExpr = Build<TKqpOpJoin>(ctx, node->Pos())
-                            .LeftInput(leftTableInput)
-                            .RightInput(rightTableInput)
-                            .JoinKind()
-                                .Value("Cross")
-                            .Build()
-                            .JoinKeys(joinKeys)
-                        .Done().Ptr();
-                        // clang-format on
-                        inputIndex += (inputIndex == 0 ? 2 : 1);
-                    }
-                } else {
-                    joinExpr = inputsInOrder.front();
-                }
-            }
-        }
-
-        filterExpr = joinExpr;
-
-        auto where = GetSetting(setItem->Tail(), "where");
-
-        if (where) {
-            TExprNode::TPtr lambda = where->Child(1)->Child(1);
-            lambda = ReplacePgOps(lambda, ctx);
-            // clang-format off
-            filterExpr = Build<TKqpOpFilter>(ctx, node->Pos())
-                .Input(filterExpr)
-                .Lambda(lambda)
-            .Done().Ptr();
-            // clang-format on
-        }
-
-        if (!filterExpr) {
-            filterExpr = Build<TKqpOpEmptySource>(ctx, node->Pos()).Done().Ptr();
-        }
-
-        // Columns to for `PgCast`
-        THashSet<TString> aggregationColumnsRequireCastToPgType;
-        // Group by fields for renames or expressions.
-        TVector<std::pair<TInfoUnit, TInfoUnit>> groupByKeysRenamesMap;
-        THashMap<uint32_t, std::pair<TInfoUnit, TExprNode::TPtr>> groupByKeysExpressionsMap;
-        // Aggregate.
-        TAggregationTraits aggTraits;
-        // Pre/Post distinct aggregations.
-        TAggregationTraits distinctAggregationTraitsPreAggregate;
-        TAggregationTraits distinctAggregationTraitsPostAggregate;
-        auto groupOps = GetSetting(setItem->Tail(), "group_exprs");
-        if (groupOps) {
-            const auto groupByList = groupOps->TailPtr();
-            for (ui32 i = 0; i < groupByList->ChildrenSize(); ++i) {
-                auto lambda = TCoLambda(ctx.DeepCopyLambda(*(groupByList->ChildPtr(i)->Child(1))));
-                auto body = lambda.Body().Ptr();
-                auto pgResolvedOp = GetPgCallable(lambda.Body().Ptr(), "PgResolvedOp");
-                // Expression for group by keys.
-                if (pgResolvedOp) {
-                    auto fromPg = ctx.NewCallable(node->Pos(), "FromPg", {pgResolvedOp});
-
-                    // clang-format off
-                    auto groupExprLambda = Build<TCoLambda>(ctx, node->Pos())
-                        .Args(lambda.Args())
-                        .Body(fromPg)
-                    .Done().Ptr();
-                    // clang-format on
-
-                    const auto newColName = TInfoUnit(GenerateUniqueColumnName("_group_expr_"));
-                    groupByKeysExpressionsMap[i] = std::make_pair(newColName, groupExprLambda);
-                    aggTraits.KeyColumns.push_back(newColName);
-                } else {
-                    TVector<TInfoUnit> keys;
-                    GetAllMembers(body, keys);
-                    Y_ENSURE(keys.size() == 1, "Invalid size of the group keys.");
-                    const auto groupKeyName = keys.front();
-                    aggTraits.KeyColumns.push_back(groupKeyName);
-                    groupByKeysRenamesMap.push_back({groupKeyName, groupKeyName});
-                }
-            }
-        }
-
-        const bool distinctAll = !!GetSetting(setItem->Tail(), "distinct_all");
-        auto result = GetSetting(setItem->Tail(), "result");
-        Y_ENSURE(result);
-        auto finalType = node->GetTypeAnn()->Cast<TListExprType>()->GetItemType()->Cast<TStructExprType>();
-
-        // Unique column names.
-        THashSet<TString> uniqueColumnNames;
-        bool needRenameMap = false;
-        // Aggregations.
-        TVector<std::pair<TInfoUnit, TInfoUnit>> aggFieldsRenamesMap;
-        TVector<std::pair<TInfoUnit, TExprNode::TPtr>> aggFieldsExpressionsMap;
-        bool distinctPreAggregate = false;
-        for (ui32 i = 0; i < result->Child(1)->ChildrenSize(); ++i) {
-            const auto resultItem = result->Child(1)->ChildPtr(i);
-            auto lambda = TCoLambda(ctx.DeepCopyLambda(*(resultItem->Child(2))));
-            auto resultColName = TString(resultItem->Child(0)->Content());
-            const auto *aggFuncResultType = finalType->FindItemType(resultColName);
-            Y_ENSURE(aggFuncResultType, "Cannot find type for aggregation result.");
-
-            auto pgAgg = GetPgCallable(lambda.Body().Ptr(), "PgAgg");
-            if (pgAgg) {
-                // Collect original column names processing `PgAgg` callable.
-                TInfoUnit originalColName;
-                TInfoUnit renamedColName;
-
-                auto lambdaBody = lambda.Body().Ptr();
-                auto pgResolvedOp = GetPgCallable(lambda.Body().Ptr(), "PgResolvedOp");
-                if (pgResolvedOp && !lambdaBody->IsCallable("PgResolvedOp")) {
-                    // Aggregation on expression f(a x b).
-                    // We pull expression outside a given aggregation and rename result of a given expression with unique name
-                    // to later process result with aggregate function.
-                    // For example: (a x b) as uqique_result_name -> f(unique_result_name)
-                    auto fromPg = ctx.NewCallable(node->Pos(), "FromPg", {pgResolvedOp});
-
-                    // clang-format off
-                    auto exprLambda = Build<TCoLambda>(ctx, node->Pos())
-                        .Args(lambda.Args())
-                        .Body(fromPg)
-                    .Done().Ptr();
-                    // clang-format on
-
-                    // Just any unique name for expression result, physical plan should be AsSturct(`unique_name (expression)).
-                    originalColName = TInfoUnit(GenerateUniqueColumnName("_expr_"));
-                    renamedColName = originalColName;
-                    aggFieldsExpressionsMap.push_back({originalColName, exprLambda});
-                } else {
-                    // Either an aggregation f(a) or expression on aggregation f(a) x b.
-                    // Here we want to get just a column name for aggregation.
-                    Y_ENSURE(pgAgg->ChildrenSize() == 3, "Invalid children size for `PgAgg`");
-                    auto toPg = pgAgg->ChildPtr(2);
-                    Y_ENSURE(toPg->IsCallable("ToPg") && toPg->ChildPtr(0)->IsCallable("Member"), "PgAgg not a member");
-                    auto member = TCoMember(toPg->ChildPtr(0));
-                    originalColName = TInfoUnit(member.Name().StringValue());
-                    renamedColName = originalColName;
-
-                    // Aggregation columns should be unique, so we have to add rename map.
-                    // For example f(a), g(a) => map((a -> a), (a -> a_0)) -> f(a), g(a_0).
-                    if (uniqueColumnNames.count(originalColName.GetFullName())) {
-                        renamedColName = TInfoUnit(originalColName.Alias, GenerateUniqueColumnName(originalColName.ColumnName));
-                        needRenameMap = true;
-                    }
-                    aggFieldsRenamesMap.push_back({originalColName, renamedColName});
-                }
-                uniqueColumnNames.insert(renamedColName.GetFullName());
-
-                // Distinct for column or expression f(distinct a) => distinct a as result -> f(result).
-                if (!!GetAtom(pgAgg->ChildPtr(1), "distinct")) {
-                    const auto colName = renamedColName.GetFullName();
-                    auto distinctAggTraits =
-                        BuildAggregationTraits(colName, colName, "distinct", aggFuncResultType, ctx, node->Pos());
-                    distinctAggregationTraitsPreAggregate.AggTraitsList.push_back(distinctAggTraits);
-                    distinctAggregationTraitsPreAggregate.KeyColumns.push_back(renamedColName);
-                    distinctPreAggregate = true;
-                }
-
-                // Aggregation on pg columns requires type cast from yql type to pg type.
-                aggregationColumnsRequireCastToPgType.insert(resultColName);
-                const TString aggFuncName = TString(pgAgg->ChildPtr(0)->Content());
-                // Build an aggregation traits.
-                auto aggregationTraits = BuildAggregationTraits(renamedColName.GetFullName(), resultColName, aggFuncName,
-                                                                aggFuncResultType, ctx, node->Pos());
-                aggTraits.AggTraitsList.push_back(aggregationTraits);
-
-                // Case for distinct after aggregation.
-                if (distinctAll) {
-                    auto distinctAggTraits =
-                        BuildAggregationTraits(resultColName, resultColName, "distinct", aggFuncResultType, ctx, node->Pos());
-                    distinctAggregationTraitsPostAggregate.AggTraitsList.push_back(distinctAggTraits);
-                    distinctAggregationTraitsPostAggregate.KeyColumns.push_back(TInfoUnit(resultColName));
-                }
-            // This case covers distinct all on just columns without aggregation functions.
-            } else if (!pgAgg && distinctAll) {
-                auto pgGroupRef = GetPgCallable(lambda.Body().Ptr(), "PgGroupRef");
-                TInfoUnit colName;
-                if (pgGroupRef) {
-                    if (pgGroupRef->ChildrenSize() == 4) {
-                        colName = TInfoUnit(TString(pgGroupRef->ChildPtr(3)->Content()));
-                    } else {
-                        Y_ENSURE(false, "Invalid column size");
-                    }
-                } else {
-                    TVector<TInfoUnit> originalColNames;
-                    GetAllMembers(resultItem->ChildPtr(2), originalColNames);
-                    Y_ENSURE(originalColNames.size() == 1, "Invalid column size for aggregation columns.");
-                    colName = originalColNames.front();
-                }
-                aggregationColumnsRequireCastToPgType.insert(colName.ColumnName);
-
-                auto distinctAggTraits =
-                    BuildAggregationTraits(colName.GetFullName(), colName.GetFullName(), "distinct", aggFuncResultType, ctx, node->Pos());
-                distinctAggregationTraitsPostAggregate.AggTraitsList.push_back(distinctAggTraits);
-                distinctAggregationTraitsPostAggregate.KeyColumns.push_back(colName);
-            }
-        }
-
-        // Distinct pre aggregate fro group by keys.
-        if (distinctPreAggregate) {
-            for (const auto& key : aggTraits.KeyColumns) {
-                const auto colName = key.GetFullName();
-                const auto* aggFuncResultType = finalType->FindItemType(key.ColumnName);
-                Y_ENSURE(aggFuncResultType, "Cannot find type for aggregation result");
-                auto distinctAggTraits = BuildAggregationTraits(colName, colName, "distinct", aggFuncResultType, ctx, node->Pos());
-                distinctAggregationTraitsPreAggregate.AggTraitsList.push_back(distinctAggTraits);
-                distinctAggregationTraitsPreAggregate.KeyColumns.push_back(colName);
-                aggregationColumnsRequireCastToPgType.insert(key.ColumnName);
-            }
-        }
-
-        // Distinct post aggregate for group by keys.
-        if (distinctAll) {
-            for (const auto& key : aggTraits.KeyColumns) {
-                const auto colName = key.GetFullName();
-                const auto* aggFuncResultType = finalType->FindItemType(key.ColumnName);
-                // agg key in result set.
-                if (aggFuncResultType) {
-                    auto distinctAggTraits = BuildAggregationTraits(colName, colName, "distinct", aggFuncResultType, ctx, node->Pos());
-                    distinctAggregationTraitsPostAggregate.AggTraitsList.push_back(distinctAggTraits);
-                    distinctAggregationTraitsPostAggregate.KeyColumns.push_back(colName);
-                    aggregationColumnsRequireCastToPgType.insert(key.ColumnName);
-                }
-            }
-        }
-
-        TExprNode::TPtr resultExpr = filterExpr;
-        // In case we have a multiple consumers for the single column we have to map and rename it.
-        if (needRenameMap) {
-            resultExpr = BuildRenameMap(resultExpr, aggFieldsRenamesMap, groupByKeysRenamesMap, ctx, node->Pos());
-        }
-        // In case we have an expression for aggregation - f(a + b ..) or group by.
-        if (!aggFieldsExpressionsMap.empty() || !groupByKeysExpressionsMap.empty()) {
-            resultExpr = BuildExpressionMap(resultExpr, aggFieldsExpressionsMap, aggFieldsRenamesMap, groupByKeysRenamesMap,
-                                            groupByKeysExpressionsMap, ctx, node->Pos());
-        }
-        // Build distinct aggregate pre aggregate.
-        if (!distinctAggregationTraitsPreAggregate.AggTraitsList.empty()) {
-            resultExpr = BuildAggregate(resultExpr, distinctAggregationTraitsPreAggregate.KeyColumns,
-                                        distinctAggregationTraitsPreAggregate.AggTraitsList, true, ctx, node->Pos());
-        }
-        // Build Aggreegate.
-        if (!aggTraits.AggTraitsList.empty()) {
-            resultExpr = BuildAggregate(resultExpr, aggTraits.KeyColumns, aggTraits.AggTraitsList, false, ctx, node->Pos());
-        }
-        // Build distinct aggregate post aggregate.
-        if (!distinctAggregationTraitsPostAggregate.AggTraitsList.empty()) {
-            resultExpr = BuildAggregate(resultExpr, distinctAggregationTraitsPostAggregate.KeyColumns,
-                                        distinctAggregationTraitsPostAggregate.AggTraitsList, true, ctx, node->Pos());
-        }
-
-        finalColumnOrder.clear();
-
-        for (auto resultItem : result->Child(1)->Children()) {
-            auto column = resultItem->Child(0);
-            TString columnName = TString(column->Content());
-
-            const auto expectedTypeNode = finalType->FindItemType(columnName);
-            Y_ENSURE(expectedTypeNode);
-            const auto expectedType = expectedTypeNode->Cast<TPgExprType>();
-            const auto actualTypeNode = resultItem->GetTypeAnn();
-
-            YQL_CLOG(TRACE, CoreDq) << "Actual type for column: " << columnName << " is: " << *actualTypeNode;
-            YQL_CLOG(TRACE, CoreDq) << "Expected type for column: " << columnName << " is: " << *expectedTypeNode;
-
-            ui32 actualPgTypeId;
-            bool convertToPg;
-            Y_ENSURE(ExtractPgType(actualTypeNode, actualPgTypeId, convertToPg, node->Pos(), ctx));
-
-            bool needPgCast = (expectedType->GetId() != actualPgTypeId);
-            auto lambda = TCoLambda(ctx.DeepCopyLambda(*(resultItem->Child(2))));
-            bool needPgCastForAgg = aggregationColumnsRequireCastToPgType.count(columnName);
-
-            auto pgAgg = GetPgCallable(lambda.Body().Ptr(), "PgAgg");
-            if (pgAgg) {
-                Y_ENSURE(SupportedAggregationFunctions.count(pgAgg->ChildPtr(0)->Content()),
-                         "Aggregation function " + TString(pgAgg->ChildPtr(0)->Content()) + " is not supported ");
-
-                // clang-format off
-                auto newBody = Build<TCoMember>(ctx, node->Pos())
-                    .Struct(lambda.Args().Arg(0))
-                    .Name<TCoAtom>()
-                        .Value(columnName)
-                    .Build()
-                .Done().Ptr();
-                // clang-format on
-
-                auto lambdaBody = lambda.Body().Ptr();
-                // Build a projection lambda, we do not need `PgAgg` inside.
-                if (lambdaBody->IsCallable("PgResolvedOp")) {
-                    // Replace PgResolvedOp(PgAgg(arg)) -> PgResolvedOp(PgCast(ToPg(Member(arg, columnName))))
-                    auto toPg = ctx.NewCallable(node->Pos(), "ToPg", {newBody});
-                    auto pgType =
-                        ctx.NewCallable(node->Pos(), "PgType", {ctx.NewAtom(node->Pos(), NPg::LookupType(expectedType->GetId()).Name)});
-                    auto pgCast = ctx.NewCallable(node->Pos(), "PgCast", {toPg, pgType});
-
-                    TNodeOnNodeOwnedMap replaces;
-                    replaces[pgAgg.Get()] = pgCast;
-                    newBody = ctx.ReplaceNodes(lambda.Body().Ptr(), replaces);
-                }
-
-                // clang-format off
-                lambda = Build<TCoLambda>(ctx, node->Pos())
-                    .Args(lambda.Args())
-                    .Body(newBody)
-                .Done();
-                // clang-format on
-            }
-
-            // Eliminate `PgGroupRef` from projection lambda.
-            auto pgGroupRef = GetPgCallable(lambda.Body().Ptr(), "PgGroupRef");
-            if (pgGroupRef) {
-                TString columnName;
-                if (pgGroupRef->ChildrenSize() == 4) {
-                    columnName = TString(pgGroupRef->ChildPtr(3)->Content());
-                } else if (pgGroupRef->ChildrenSize() == 3) {
-                    // In this case we can get a column name from group expr map
-                    const auto groupByKeyExprId = FromString<uint32_t>(TString(pgGroupRef->ChildPtr(2)->Content()));
-                    auto it = groupByKeysExpressionsMap.find(groupByKeyExprId);
-                    Y_ENSURE(it != groupByKeysExpressionsMap.end(), "Group by key expression has invalid content.");
-                    columnName = it->second.first.GetFullName();
-                    // Always need a pg cast for expressions.
-                    needPgCast = true;
-                } else {
-                    Y_ENSURE(false, "Invalid children size for `pgGroupRef`");
-                }
-
-                // clang-format off
-                lambda = Build<TCoLambda>(ctx, node->Pos())
-                    .Args(lambda.Args())
-                    .Body<TCoMember>()
-                        .Struct(lambda.Args().Arg(0))
-                        .Name<TCoAtom>()
-                            .Value(columnName)
-                        .Build()
-                    .Build()
-                .Done();
-                // clang-format on
-            }
-
-            if (convertToPg && !needPgCastForAgg) {
-                Y_ENSURE(!needPgCast,
-                         TStringBuilder() << "Conversion to PG type is different at typization (" << expectedType->GetId()
-                                          << ") and optimization (" << actualPgTypeId << ") stages.");
-
-                TExprNode::TPtr lambdaBody = lambda.Body().Ptr();
-                lambdaBody = ReplacePgOps(lambdaBody, ctx);
-                auto toPg = ctx.NewCallable(node->Pos(), "ToPg", {lambdaBody});
-
-                // clang-format off
-                lambda = Build<TCoLambda>(ctx, node->Pos())
-                    .Args(lambda.Args())
-                    .Body(toPg)
-                .Done();
-                // clang-format on
-            } else if ((needPgCast || needPgCastForAgg)) {
-
-                auto pgType =
-                    ctx.NewCallable(node->Pos(), "PgType", {ctx.NewAtom(node->Pos(), NPg::LookupType(expectedType->GetId()).Name)});
-                TExprNode::TPtr lambdaBody = lambda.Body().Ptr();
-                lambdaBody = ReplacePgOps(lambdaBody, ctx);
-                auto pgCast = ctx.NewCallable(node->Pos(), "PgCast", {lambdaBody, pgType});
-
-                // clang-format off
-                lambda = Build<TCoLambda>(ctx, node->Pos())
-                    .Args(lambda.Args())
-                    .Body(pgCast)
-                .Done();
-                // clang-format on
-            }
-
-            if (resultElementCounters.contains(columnName)) {
-                resultElementCounters[columnName] += 1;
-                columnName = columnName + "_generated_" + std::to_string(resultElementCounters.at(columnName));
-            } else {
-                resultElementCounters[columnName] = 1;
-            }
-
-            finalColumnOrder.push_back(columnName);
-            auto variable = Build<TCoAtom>(ctx, node->Pos()).Value(columnName).Done();
-
-            // clang-format off
-            resultElements.push_back(Build<TKqpOpMapElementLambda>(ctx, node->Pos())
-                .Input(resultExpr)
-                .Variable(variable)
-                .Lambda(lambda)
-            .Done().Ptr());
-            // clang-format on
-        }
-
-        // clang-format off
-        auto setItemPtr = Build<TKqpOpMap>(ctx, node->Pos())
-            .Input(resultExpr)
-            .MapElements()
-                .Add(resultElements)
-            .Build()
-            .Project()
-                .Value("true")
-            .Build()
-        .Done().Ptr();
-        // clang-format onto
-
-        auto sort = GetSetting(setItem->Tail(), "sort");
-        if (sort) {
-            setItemPtr = BuildSort(setItemPtr, sort, ctx);
-        }
-
-        setItemsResults.push_back(setItemPtr);
-    }
-
-    auto setOps = GetSetting(node->Head(), "set_ops");
-    Y_ENSURE(setOps && setItemsResults.size());
-
-    auto setOpsList = setOps->TailPtr();
-    TExprNode::TPtr opResult = setItemsResults.front();
-    for (ui32 i = 0, end = setOpsList->ChildrenSize(), setItemsIndex = 0, opsInputCount = 0; i < end; ++i) {
-        if (setOpsList->ChildPtr(i)->Content() == "push") {
-            ++opsInputCount;
-            continue;
-        }
-        Y_ENSURE(setOpsList->ChildPtr(i)->Content() == "union_all");
-        Y_ENSURE(opsInputCount <= 2);
-
-        TExprNode::TPtr leftInput;
-        TExprNode::TPtr rightInput;
-        if (opsInputCount == 2) {
-            Y_ENSURE(setItemsIndex + 1 < end);
-            leftInput = setItemsResults[setItemsIndex++];
-            rightInput = setItemsResults[setItemsIndex++];
-        } else {
-            Y_ENSURE(setItemsIndex < end);
-            leftInput = opResult;
-            rightInput = setItemsResults[setItemsIndex++];
-        }
-
-        // clang-format off
-        opResult = Build<TKqpOpUnionAll>(ctx, node->Pos())
-            .LeftInput(leftInput)
-            .RightInput(rightInput)
-        .Done().Ptr();
-        // clang-format on
-
-        // Count again.
-        opsInputCount = 0;
-    }
-
-    auto sort = GetSetting(node->Head(), "sort");
-    if (sort) {
-        opResult = BuildSort(opResult, sort, ctx);
-    }
-
-    TVector<TCoAtom> columnAtomList;
-    for (auto c : finalColumnOrder) {
-        columnAtomList.push_back(Build<TCoAtom>(ctx, node->Pos()).Value(c).Done());
-    }
-    auto columnOrder = Build<TCoAtomList>(ctx, node->Pos()).Add(columnAtomList).Done().Ptr();
-
-    // clang-format off
-    return Build<TKqpOpRoot>(ctx, node->Pos())
-        .Input(opResult)
-        .ColumnOrder(columnOrder)
-    .Done().Ptr();
-    // clang-format on
-
-}
 
 TExprNode::TPtr PushTakeIntoPlan(const TExprNode::TPtr &node, TExprContext &ctx, const TTypeAnnotationContext &typeCtx) {
     Y_UNUSED(typeCtx);
@@ -1006,6 +22,7 @@ TExprNode::TPtr PushTakeIntoPlan(const TExprNode::TPtr &node, TExprContext &ctx,
                 .Count(take.Count())
             .Build()
             .ColumnOrder(root.Cast().ColumnOrder())
+            .PgSyntax(root.Cast().PgSyntax())
         .Done().Ptr();
         // clang-format on
     } else {
@@ -1013,23 +30,23 @@ TExprNode::TPtr PushTakeIntoPlan(const TExprNode::TPtr &node, TExprContext &ctx,
     }
 }
 
-TExprNode::TPtr RewritePgSublink(const TExprNode::TPtr &node, TExprContext &ctx) {
+TExprNode::TPtr RewriteSublink(const TExprNode::TPtr &node, TExprContext &ctx) {
     if (node->Child(0)->Content() != "expr") {
         return node;
     }
 
     // clang-format off
-    return Build<TKqpPgExprSublink>(ctx, node->Pos())
+    return Build<TKqpExprSublink>(ctx, node->Pos())
         .Expr(node->Child(4))
         .Done().Ptr();
     // clang-format on
 }
 
 TExprNode::TPtr RemoveRootFromSublink(const TExprNode::TPtr &node, TExprContext &ctx) {
-    auto sublink = TKqpPgExprSublink(node);
+    auto sublink = TKqpExprSublink(node);
     if (auto root = sublink.Expr().Maybe<TKqpOpRoot>()) {
         // clang-format off
-        return Build<TKqpPgExprSublink>(ctx, node->Pos())
+        return Build<TKqpExprSublink>(ctx, node->Pos())
             .Expr(root.Cast().Input())
             .Done().Ptr();
         // clang-format on
@@ -1041,7 +58,7 @@ TExprNode::TPtr RemoveRootFromSublink(const TExprNode::TPtr &node, TExprContext 
 namespace NKikimr {
 namespace NKqp {
 
-IGraphTransformer::TStatus TKqpPgRewriteTransformer::DoTransform(TExprNode::TPtr input, TExprNode::TPtr &output, TExprContext &ctx) {
+IGraphTransformer::TStatus TKqpRewriteSelectTransformer::DoTransform(TExprNode::TPtr input, TExprNode::TPtr &output, TExprContext &ctx) {
     output = input;
     TOptimizeExprSettings settings(&TypeCtx);
 
@@ -1049,7 +66,9 @@ IGraphTransformer::TStatus TKqpPgRewriteTransformer::DoTransform(TExprNode::TPtr
         output, output,
         [](const TExprNode::TPtr &node, TExprContext &ctx) -> TExprNode::TPtr {
             if (node->IsCallable("PgSubLink")) {
-                return RewritePgSublink(node, ctx);
+                return RewriteSublink(node, ctx);
+            } else if (node->IsCallable("YqlSubLink")) {
+                return RewriteSublink(node, ctx);
             } else {
                 return node;
             }
@@ -1063,11 +82,18 @@ IGraphTransformer::TStatus TKqpPgRewriteTransformer::DoTransform(TExprNode::TPtr
     status = OptimizeExpr(
         output, output,
         [this](const TExprNode::TPtr &node, TExprContext &ctx) -> TExprNode::TPtr {
+
+            // PostgreSQL AST rewrtiting
             if (TCoPgSelect::Match(node.Get())) {
-                return RewritePgSelect(node, ctx, TypeCtx);
-            } else if (TKqpPgExprSublink::Match(node.Get())) {
+                return RewriteSelect(node, ctx, TypeCtx, true);
+            }
+            
+            // YQL AST rewriting
+            else if (TCoYqlSelect::Match(node.Get())) {
+                return RewriteSelect(node, ctx, TypeCtx, false);
+            } else if (TKqpExprSublink::Match(node.Get())) {
                 return RemoveRootFromSublink(node, ctx);
-            } else if (TCoTake::Match(node.Get())) {
+            }  else if (TCoTake::Match(node.Get())) {
                 return PushTakeIntoPlan(node, ctx, TypeCtx);
             } else {
                 return node;
@@ -1078,7 +104,7 @@ IGraphTransformer::TStatus TKqpPgRewriteTransformer::DoTransform(TExprNode::TPtr
     return status;
 }
 
-void TKqpPgRewriteTransformer::Rewind() {}
+void TKqpRewriteSelectTransformer::Rewind() {}
 
 IGraphTransformer::TStatus TKqpNewRBOTransformer::DoTransform(TExprNode::TPtr input, TExprNode::TPtr &output, TExprContext &ctx) {
     output = input;
@@ -1145,9 +171,9 @@ IGraphTransformer::TStatus TKqpRBOCleanupTransformer::DoTransform(TExprNode::TPt
 
 void TKqpRBOCleanupTransformer::Rewind() {}
 
-TAutoPtr<IGraphTransformer> CreateKqpPgRewriteTransformer(const TIntrusivePtr<TKqpOptimizeContext> &kqpCtx,
+TAutoPtr<IGraphTransformer> CreateKqpRewriteSelectTransformer(const TIntrusivePtr<TKqpOptimizeContext> &kqpCtx,
                                                           TTypeAnnotationContext &typeCtx) {
-    return new TKqpPgRewriteTransformer(kqpCtx, typeCtx);
+    return new TKqpRewriteSelectTransformer(kqpCtx, typeCtx);
 }
 
 TAutoPtr<IGraphTransformer> CreateKqpNewRBOTransformer(const TIntrusivePtr<TKqpOptimizeContext> &kqpCtx, TTypeAnnotationContext &typeCtx,
