+++ ydb/core/kqp/ut/rbo/kqp_rbo_ut.cpp
@@ -188,14 +188,13 @@ Y_UNIT_TEST_SUITE(KqpRbo) {
 
         std::vector<std::string> queries = {
             R"(
-                PRAGMA TablePathPrefix='/Root';
                 PRAGMA YqlSelect = 'force';
-                SELECT id as id2 FROM foo WHERE name != '3_name';
+                SELECT id as id2 FROM `/Root/foo` WHERE name != '3_name';
             )"
         };
 
         std::vector<std::string> results = {
-            R"([["0"];["1"];["2"];["4"];["5"];["6"];["7"];["8"];["9"]])",
+            R"([[0];[1];[2];[4];[5];[6];[7];[8];[9]])",
         };
 
         for (ui32 i = 0; i < queries.size(); ++i) {
